+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   been removed ([RHEL-49521])
 - Following pacemaker 3, support for multiple rules in a single location
   constraint has been removed ([RHEL-62719])
+- Removed pcsd configuration option PCSD_DISABLE_GUI. It is possible not to
+  install webui support by running `./configure` without flag `--enable-webui`.
 
 ### Added
 - Support for output formats `json` and `cmd` to `pcs tag config` command